@@ -4,15 +4,12 @@
   "description": "React mentions input",
   "main": "lib/index.js",
   "scripts": {
-    "test": "script/test",
-    "transpile": "babel src --out-dir lib",
-    "dist": "script/dist",
-    "start": "cd gh-pages && webpack-dev-server --config webpack.hot.config.js --port 9090 --d --hot --inline --progress --colors",
-    "bump-patch": "mversion patch -m",
-    "bump-minor": "mversion minor -m",
-    "bump-major": "mversion major -m",
-    "predist": "npm run transpile",
-    "pretranspile": "npm run test",
+    "test": "NODE_ENV=test mocha",
+    "transpile": "NODE_ENV=production babel src --out-dir lib",
+    "clean": "rimraf lib",
+    "start": "cd gh-pages && webpack-dev-server --config webpack.dev.config.js --port 9090 --d --hot --inline --progress --colors",
+    "pretranspile": "npm run clean",
+    "prepublish": "npm run transpile",
     "prepages-publish": "cd gh-pages && webpack",
     "pages-publish": "grunt publish"
   },
@@ -32,55 +29,50 @@
   },
   "homepage": "https://github.com/effektif/react-mentions",
   "devDependencies": {
-    "babel": "^5.8.29",
-    "babel-core": "^5.8.29",
-    "babel-loader": "^5.3.2",
-    "babelify": "^6.1.1",
-    "browserify": "4.2.3",
-    "browserify-shim": "3.6.0",
+    "babel": "^6.5.1",
+    "babel-cli": "^6.5.1",
+    "babel-core": "^6.5.1",
+    "babel-loader": "^6.2.2",
+    "babel-plugin-react-transform": "^2.0.0",
+    "babel-plugin-transform-class-properties": "^6.5.0",
+    "babel-plugin-transform-runtime": "^6.5.0",
+    "babel-preset-es2015": "^6.5.0",
+    "babel-preset-react": "^6.5.0",
+    "babel-preset-stage-0": "^6.5.0",
     "bundle-loader": "0.5.1",
-    "chai": "1.10.0",
+    "chai": "^3.0.0",
+    "chai-enzyme": "^0.4.1",
+    "chai-things": "^0.2.0",
     "css-loader": "^0.20.0",
     "envify": "1.2.0",
+    "enzyme": "^2.0.0",
     "grunt": ">=0.4.1",
     "grunt-contrib-copy": ">=0.4.1",
     "grunt-contrib-less": "~1.0.0",
     "grunt-gh-pages": "~0.9.1",
-    "karma": "^0.13.11",
-    "karma-babel-preprocessor": "^5.2.2",
-    "karma-browserify": "^4.4.0",
-    "karma-chrome-launcher": "0.1.4",
-    "karma-cli": "0.0.4",
-    "karma-coverage": "~0.2.7",
-    "karma-firefox-launcher": "0.1.3",
-    "karma-junit-reporter": "~0.2.2",
-    "karma-mocha": "0.1.3",
-    "karma-phantomjs-launcher": "~0.1.4",
+    "isparta": "^4.0.0",
+    "jquery": "^2.2.1",
+    "jsdom": "^8.0.4",
     "less": "~2.3.1",
     "less-loader": "^2.2.0",
-    "mocha": "1.20.1",
+    "mocha": "^1.20.1",
     "mversion": "^1.10.0",
-    "react": "~0.14.0",
-    "react-hot-loader": "^1.2.5",
+    "react-addons-test-utils": "^0.14.7",
     "react-router": "^1.0.0-rc3",
     "react-tools": "~0.14.0-alpha3",
-    "reactify": "0.15.x",
-    "rf-release": "0.3.2",
+    "react-transform-hmr": "^1.0.2",
+    "rimraf": "^2.5.2",
     "sinon": "~1.12.1",
-    "sinon-chai": "~2.6.0",
+    "sinon-chai": "^2.8.0",
     "style-loader": "^0.12.4",
-    "uglify-js": "2.4.15",
-    "webpack": "1.4.5",
-    "webpack-dev-server": "1.6.5"
-  },
-  "peerDependencies": {
-    "react": ">=0.14.0"
-  },
-  "browserify-shim": {
-    "react": "global:React",
-    "react/addons": "global:React"
+    "webpack": "^1.12.13",
+    "webpack-dev-server": "^1.14.1"
   },
   "dependencies": {
-    "fbjs": "^0.4.0"
+    "babel-runtime": "^6.5.0",
+    "lodash": "^4.5.1",
+    "react": ">=0.14.0",
+    "react-dom": ">=0.14.0",
+    "substyle": "^1.1.1"
   }
 }
