@@ -1,15 +1,14 @@
-var React = require("react");
-var ReactMentions = require("react-mentions");
+import React from "react";
+import { Mention, MentionsInput } from "react-mentions";
 
-var MentionsMixin = require("../mixins/MentionsMixin");
+import MentionsMixin from "../mixins/MentionsMixin";
 
-
-var MentionsInput = ReactMentions.MentionsInput;
-var Mention = ReactMentions.Mention;
+import defaultStyle from "./defaultStyle";
+import defaultMentionStyle from "./defaultMentionStyle";
 
 // use first/outer capture group to extract the full entered sequence to be replaced
 // and second/inner capture group to extract search string from the match
-var emailRegex = /(([^\s@]+@[^\s@]+\.[^\s@]+))$/;
+const emailRegex = /(([^\s@]+@[^\s@]+\.[^\s@]+))$/;
 
 module.exports = React.createClass({
 
@@ -32,6 +31,7 @@ module.exports = React.createClass({
         <MentionsInput
           value={this.state.value}
           onChange={this.handleChange}
+          style={ defaultStyle() }
           markup="@[__display__](__type__:__id__)"
           placeholder={"Mention people using '@'"}>
 
@@ -41,12 +41,12 @@ module.exports = React.createClass({
             data={ this.props.data }
             renderSuggestion={this.renderSuggestion}
             onAdd={this.handleAdd}
-            onRemove={this.handleRemove} />
+            onRemove={this.handleRemove} style={defaultMentionStyle} />
           <Mention
             type="email"
             trigger={emailRegex}
             data={this.requestEmail}
-            onAdd={this.handleEmailAdd} />
+            onAdd={this.handleEmailAdd} style={{ backgroundColor: '#d1c4e9'}}/>
         </MentionsInput>
       </div>
     );
@@ -60,7 +60,7 @@ module.exports = React.createClass({
     console.log("added a new mention", arguments);
   },
 
-  renderSuggestion: function(id, display, search, highlightedDisplay) {
+  renderSuggestion: function(suggestion, search, highlightedDisplay) {
     return (
       <div className="user">
         { highlightedDisplay }
