@@ -1,10 +1,8 @@
-"use strict";
-
 var PLACEHOLDERS = {
   id: "__id__",
   display: "__display__",
   type: "__type__"
-};
+}
 
 var escapeMap = {
   '&': '&amp;',
@@ -14,24 +12,24 @@ var escapeMap = {
   "'": '&#x27;',
   '`': '&#x60;'
 };
-var createEscaper = function createEscaper(map) {
-  var escaper = function escaper(match) {
+var createEscaper = function(map) {
+  var escaper = function(match) {
     return map[match];
   };
   var keys = [];
-  for (var key in map) {
-    if (map.hasOwnProperty(key)) keys.push(key);
+  for(var key in map) {
+    if(map.hasOwnProperty(key)) keys.push(key);
   }
   var source = '(?:' + keys.join('|') + ')';
   var testRegexp = RegExp(source);
   var replaceRegexp = RegExp(source, 'g');
-  return function (string) {
+  return function(string) {
     string = string == null ? '' : '' + string;
     return testRegexp.test(string) ? string.replace(replaceRegexp, escaper) : string;
   };
 };
 
-var numericComparator = function numericComparator(a, b) {
+var numericComparator = function(a, b) {
   a = a === null ? Number.MAX_VALUE : a;
   b = b === null ? Number.MAX_VALUE : b;
   return a - b;
@@ -41,48 +39,48 @@ module.exports = {
 
   escapeHtml: createEscaper(escapeMap),
 
-  escapeRegex: function escapeRegex(str) {
-    return str.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
+  escapeRegex: function(str) {
+      return str.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
   },
 
-  markupToRegex: function markupToRegex(markup, matchAtEnd) {
+  markupToRegex: function(markup, matchAtEnd) {
     var markupPattern = this.escapeRegex(markup);
     markupPattern = markupPattern.replace(PLACEHOLDERS.display, "(.+?)");
     markupPattern = markupPattern.replace(PLACEHOLDERS.id, "(.+?)");
     markupPattern = markupPattern.replace(PLACEHOLDERS.type, "(.+?)");
-    if (matchAtEnd) {
+    if(matchAtEnd) {
       // append a $ to match at the end of the string
       markupPattern = markupPattern + "$";
     }
     return new RegExp(markupPattern, "g");
   },
 
-  spliceString: function spliceString(str, start, end, insert) {
+  spliceString: function(str, start, end, insert) {
     return str.substring(0, start) + insert + str.substring(end);
   },
 
-  extend: function extend(obj) {
+  extend: function(obj) {
     var source, prop;
     for (var i = 1, length = arguments.length; i < length; i++) {
       source = arguments[i];
       for (prop in source) {
         if (hasOwnProperty.call(source, prop)) {
-          obj[prop] = source[prop];
+            obj[prop] = source[prop];
         }
       }
     }
     return obj;
   },
 
-  isNumber: function isNumber(obj) {
+  isNumber: function(obj) {
     return Object.prototype.toString.call(obj) === "[object Number]";
   },
 
   /**
    * parameterName: "id", "display", or "type"
    */
-  getPositionOfCapturingGroup: function getPositionOfCapturingGroup(markup, parameterName) {
-    if (parameterName !== "id" && parameterName !== "display" && parameterName !== "type") {
+  getPositionOfCapturingGroup: function(markup, parameterName) {
+    if(parameterName !== "id" && parameterName !== "display" && parameterName !== "type") {
       throw new Error("parameterName must be 'id', 'display', or 'type'");
     }
 
@@ -92,16 +90,16 @@ module.exports = {
     var indexType = markup.indexOf(PLACEHOLDERS.type);
 
     // set indices to null if not found
-    if (indexDisplay < 0) indexDisplay = null;
-    if (indexId < 0) indexId = null;
-    if (indexType < 0) indexType = null;
+    if(indexDisplay < 0) indexDisplay = null;
+    if(indexId < 0) indexId = null;
+    if(indexType < 0) indexType = null;
 
-    if (indexDisplay === null && indexId === null) {
+    if(indexDisplay === null && indexId === null) {
       // markup contains none of the mandatory placeholders
       throw new Error("The markup `" + markup + "` must contain at least one of the placeholders `__id__` or `__display__`");
     }
 
-    if (indexType === null && parameterName === "type") {
+    if(indexType === null && parameterName === "type") {
       // markup does not contain optional __type__ placeholder
       return null;
     }
@@ -111,18 +109,19 @@ module.exports = {
 
     // If only one the placeholders __id__ and __display__ is present,
     // use the captured string for both parameters, id and display
-    if (indexDisplay === null) indexDisplay = indexId;
-    if (indexId === null) indexId = indexDisplay;
+    if(indexDisplay === null) indexDisplay = indexId;
+    if(indexId === null) indexId = indexDisplay;
 
-    if (parameterName === "id") return sortedIndices.indexOf(indexId);
-    if (parameterName === "display") return sortedIndices.indexOf(indexDisplay);
-    if (parameterName === "type") return indexType === null ? null : sortedIndices.indexOf(indexType);
+    if(parameterName === "id") return sortedIndices.indexOf(indexId);
+    if(parameterName === "display") return sortedIndices.indexOf(indexDisplay);
+    if(parameterName === "type") return indexType === null ? null : sortedIndices.indexOf(indexType);
+
   },
 
   // Finds all occurences of the markup in the value and iterates the plain text sub strings
   // in between those markups using `textIteratee` and the markup occurrences using the
   // `markupIteratee`.
-  iterateMentionsMarkup: function iterateMentionsMarkup(value, markup, textIteratee, markupIteratee, displayTransform) {
+  iterateMentionsMarkup: function(value, markup, textIteratee, markupIteratee, displayTransform) {
     var regex = this.markupToRegex(markup);
     var displayPos = this.getPositionOfCapturingGroup(markup, "display");
     var idPos = this.getPositionOfCapturingGroup(markup, "id");
@@ -133,26 +132,26 @@ module.exports = {
     var currentPlainTextIndex = 0;
 
     // detect all mention markup occurences in the value and iterate the matches
-    while ((match = regex.exec(value)) !== null) {
+    while((match = regex.exec(value)) !== null) {
 
-      var id = match[idPos + 1];
-      var display = match[displayPos + 1];
-      var type = typePos !== null ? match[typePos + 1] : null;
+      var id = match[idPos+1];
+      var display = match[displayPos+1];
+      var type = typePos !== null ? match[typePos+1] : null;
 
-      if (displayTransform) display = displayTransform(id, display, type);
+      if(displayTransform) display = displayTransform(id, display, type);
 
       var substr = value.substring(start, match.index);
-      textIteratee(substr, start, currentPlainTextIndex);
+      textIteratee( substr, start, currentPlainTextIndex );
       currentPlainTextIndex += substr.length;
 
-      markupIteratee(match[0], match.index, currentPlainTextIndex, id, display, type, start);
+      markupIteratee( match[0], match.index, currentPlainTextIndex, id, display, type, start );
       currentPlainTextIndex += display.length;
 
       start = regex.lastIndex;
     }
 
-    if (start < value.length) {
-      textIteratee(value.substring(start), start, currentPlainTextIndex);
+    if(start < value.length) {
+      textIteratee( value.substring(start), start, currentPlainTextIndex );
     }
   },
 
@@ -160,24 +159,24 @@ module.exports = {
   // in the marked up value string.
   // If the passed character index lies inside a mention, returns the index of the mention
   // markup's first char, or respectively tho one after its last char, if the flag `toEndOfMarkup` is set.
-  mapPlainTextIndex: function mapPlainTextIndex(value, markup, indexInPlainText, toEndOfMarkup, displayTransform) {
-    if (!this.isNumber(indexInPlainText)) {
+  mapPlainTextIndex: function(value, markup, indexInPlainText, toEndOfMarkup, displayTransform) {
+    if(!this.isNumber(indexInPlainText)) {
       return indexInPlainText;
     }
 
     var result;
-    var textIteratee = function textIteratee(substr, index, substrPlainTextIndex) {
-      if (result !== undefined) return;
+    var textIteratee = function(substr, index, substrPlainTextIndex) {
+      if(result !== undefined) return;
 
-      if (substrPlainTextIndex + substr.length >= indexInPlainText) {
+      if(substrPlainTextIndex + substr.length >= indexInPlainText) {
         // found the corresponding position in the current plain text range
         result = index + indexInPlainText - substrPlainTextIndex;
       }
     };
-    var markupIteratee = function markupIteratee(markup, index, mentionPlainTextIndex, id, display, type, lastMentionEndIndex) {
-      if (result !== undefined) return;
+    var markupIteratee = function(markup, index, mentionPlainTextIndex, id, display, type, lastMentionEndIndex) {
+      if(result !== undefined) return;
 
-      if (mentionPlainTextIndex + display.length > indexInPlainText) {
+      if(mentionPlainTextIndex + display.length > indexInPlainText) {
         // found the corresponding position inside current match,
         // return the index of the first or after the last char of the matching markup
         // depending on whether the `toEndOfMarkup` is set
@@ -195,17 +194,17 @@ module.exports = {
   // For a given indexInPlainText that lies inside a mention,
   // returns a the index of of the first char of the mention in the plain text.
   // If indexInPlainText does not lie inside a mention, returns indexInPlainText.
-  findStartOfMentionInPlainText: function findStartOfMentionInPlainText(value, markup, indexInPlainText, displayTransform) {
+  findStartOfMentionInPlainText: function(value, markup, indexInPlainText, displayTransform) {
     var result = indexInPlainText;
     var foundMention = false;
 
-    var markupIteratee = function markupIteratee(markup, index, mentionPlainTextIndex, id, display, type, lastMentionEndIndex) {
-      if (mentionPlainTextIndex < indexInPlainText && mentionPlainTextIndex + display.length > indexInPlainText) {
+    var markupIteratee = function(markup, index, mentionPlainTextIndex, id, display, type, lastMentionEndIndex) {
+      if(mentionPlainTextIndex < indexInPlainText && mentionPlainTextIndex + display.length > indexInPlainText) {
         result = mentionPlainTextIndex;
         foundMention = true;
       }
     };
-    this.iterateMentionsMarkup(value, markup, function () {}, markupIteratee, displayTransform);
+    this.iterateMentionsMarkup(value, markup, function(){}, markupIteratee, displayTransform);
 
     if (foundMention) {
       return result;
@@ -213,14 +212,14 @@ module.exports = {
   },
 
   // Returns whether the given plain text index lies inside a mention
-  isInsideOfMention: function isInsideOfMention(value, markup, indexInPlainText, displayTransform) {
+  isInsideOfMention: function(value, markup, indexInPlainText, displayTransform) {
     var mentionStart = this.findStartOfMentionInPlainText(value, markup, indexInPlainText, displayTransform);
-    return mentionStart !== undefined && mentionStart !== indexInPlainText;
+    return mentionStart !== undefined && mentionStart !== indexInPlainText
   },
 
   // Applies a change from the plain text textarea to the underlying marked up value
   // guided by the textarea text selection ranges before and after the change
-  applyChangeToValue: function applyChangeToValue(value, markup, plainTextValue, selectionStartBeforeChange, selectionEndBeforeChange, selectionEndAfterChange, displayTransform) {
+  applyChangeToValue: function(value, markup, plainTextValue, selectionStartBeforeChange, selectionEndBeforeChange, selectionEndAfterChange, displayTransform) {
     var oldPlainTextValue = this.getPlainText(value, markup, displayTransform);
 
     var lengthDelta = oldPlainTextValue.length - plainTextValue.length;
@@ -233,7 +232,10 @@ module.exports = {
     }
 
     // Fixes an issue with replacing combined characters for complex input. Eg like acented letters on OSX
-    if (selectionStartBeforeChange === selectionEndBeforeChange && selectionEndBeforeChange === selectionEndAfterChange && oldPlainTextValue.length === plainTextValue.length) {
+    if (selectionStartBeforeChange === selectionEndBeforeChange &&
+      selectionEndBeforeChange === selectionEndAfterChange &&
+      oldPlainTextValue.length === plainTextValue.length
+    ) {
       selectionStartBeforeChange = selectionStartBeforeChange - 1;
     }
 
@@ -244,33 +246,38 @@ module.exports = {
     var spliceStart = Math.min(selectionStartBeforeChange, selectionEndAfterChange);
 
     var spliceEnd = selectionEndBeforeChange;
-    if (selectionStartBeforeChange === selectionEndAfterChange) {
+    if(selectionStartBeforeChange === selectionEndAfterChange) {
       // handling for Delete key with no range selection
       spliceEnd = Math.max(selectionEndBeforeChange, selectionStartBeforeChange + lengthDelta);
     }
 
     // splice the current marked up value and insert new chars
-    return this.spliceString(value, this.mapPlainTextIndex(value, markup, spliceStart, false, displayTransform), this.mapPlainTextIndex(value, markup, spliceEnd, true, displayTransform), insert);
+    return this.spliceString(
+      value,
+      this.mapPlainTextIndex(value, markup, spliceStart, false, displayTransform),
+      this.mapPlainTextIndex(value, markup, spliceEnd, true, displayTransform),
+      insert
+    );
   },
 
-  getPlainText: function getPlainText(value, markup, displayTransform) {
+  getPlainText: function(value, markup, displayTransform) {
     var regex = this.markupToRegex(markup);
     var idPos = this.getPositionOfCapturingGroup(markup, "id");
     var displayPos = this.getPositionOfCapturingGroup(markup, "display");
     var typePos = this.getPositionOfCapturingGroup(markup, "type");
-    return value.replace(regex, function () {
+    return value.replace(regex, function() {
       // first argument is the whole match, capturing groups are following
-      var id = arguments[idPos + 1];
-      var display = arguments[displayPos + 1];
-      var type = arguments[typePos + 1];
-      if (displayTransform) display = displayTransform(id, display, type);
+      var id = arguments[idPos+1];
+      var display = arguments[displayPos+1];
+      var type = arguments[typePos+1];
+      if(displayTransform) display = displayTransform(id, display, type);
       return display;
     });
   },
 
-  getMentions: function getMentions(value, markup) {
+  getMentions: function (value, markup) {
     var mentions = [];
-    this.iterateMentionsMarkup(value, markup, function () {}, function (match, index, plainTextIndex, id, display, type, start) {
+    this.iterateMentionsMarkup(value, markup, function (){}, function (match, index, plainTextIndex, id, display, type, start) {
       mentions.push({
         id: id,
         display: display,
@@ -282,11 +289,49 @@ module.exports = {
     return mentions;
   },
 
-  makeMentionsMarkup: function makeMentionsMarkup(markup, id, display, type) {
+  makeMentionsMarkup: function(markup, id, display, type) {
     var result = markup.replace(PLACEHOLDERS.id, id);
     result = result.replace(PLACEHOLDERS.display, display);
     result = result.replace(PLACEHOLDERS.type, type);
     return result;
+  },
+
+  countSuggestions: function(suggestions) {
+    let result = 0;
+    for(let prop in suggestions) {
+      if(suggestions.hasOwnProperty(prop)) {
+        result += suggestions[prop].results.length;
+      }
+    }
+    return result;
+  },
+
+  getSuggestions: function(suggestions) {
+    var result = [];
+
+    for(var mentionType in suggestions) {
+      if(!suggestions.hasOwnProperty(mentionType)) {
+        return;
+      }
+
+      result = result.concat({
+        suggestions: suggestions[mentionType].results,
+        descriptor: suggestions[mentionType]
+      });
+    }
+
+    return result;
+  },
+
+  getSuggestion: function(suggestions, index) {
+    return this.getSuggestions(suggestions).reduce((result, { suggestions, descriptor }) => [
+      ...result,
+
+      ...suggestions.map((suggestion) => ({
+        suggestion: suggestion,
+        descriptor: descriptor
+      }))
+    ], [])[index];
   }
 
-};
+}
